appState renamed to appUiState
